Label_Studio/requirements.txt
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.11.0
pillow>=10.0.0
opencv-python>=4.9.0
python-dotenv>=1.0.0
openai>=1.0.0


